NP-3206 Generalize period input fields
